+++ src/layers/ImageLayer.js
@@ -73,7 +73,10 @@ const defaultProps = {
  */
 const ImageLayer = class extends CompositeLayer {
   initializeState() {
+    const internalState = this.internalState;
+    const handler = () => onPointer(internalState.layer);
     this.state = {
+      onPointer: handler,
       unprojectLensBounds: [0, 0, 0, 0],
       width: 0,
       height: 0,
@@ -81,15 +84,23 @@ const ImageLayer = class extends CompositeLayer {
     };
     if (this.context.deck) {
       this.context.deck.eventManager.on({
-        pointermove: () => onPointer(this),
-        pointerleave: () => onPointer(this),
-        wheel: () => onPointer(this)
+        pointermove: handler,
+        pointerleave: handler,
+        wheel: handler
       });
     }
   }
 
   finalizeState() {
     this.state.abortController.abort();
+    if (this.context.deck) {
+      this.context.deck.eventManager.off({
+        pointermove: this.state.onPointer,
+        pointerleave: this.state.onPointer,
+        wheel: this.state.onPointer
+      });
+    }
+    super.finalizeState();
   }
 
   updateState({ props, oldProps }) {